refactor(sync): add lock_for_init helper and centralize init-time best-effort recovery

+++ drivers/nvme/src/controller.rs
@@ -259,8 +259,8 @@ impl NvmeCapabilities {
     }
 
     /// 最大キュー深度
-    pub fn max_queue_depth(&self) -> u16 {
-        self.mqes() + 1
+    pub fn max_queue_depth(&self) -> u32 {
+        (self.mqes() as u32) + 1
     }
 }
